added a custom characteristic, uuid generated

+++ src/lib/messages.rs
@@ -1,7 +1,8 @@
 use embassy_sync::blocking_mutex::raw::CriticalSectionRawMutex;
 use embassy_sync::signal::Signal;
 
-pub static ADC_SIGNAL: Signal<CriticalSectionRawMutex, u8> = Signal::new();
+//pub static ADC_SIGNAL: Signal<CriticalSectionRawMutex, u8> = Signal::new();
+pub static ADC_SIGNAL: Signal<CriticalSectionRawMutex, i16> = Signal::new();
 
 pub struct Enviro {
     pub temperature: i16,
